requirements.txt
awsebcli==3.7.7
blessed==1.9.5
botocore==1.4.36
cement==2.8.2
colorama==0.3.7
Django==1.9.6
django-allauth==0.26.1
django-braces==1.9.0
django-oauth-toolkit==0.10.0
django-rest-auth==0.7.0
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
httpie==0.9.3
jmespath==0.9.0
MySQL-python==1.2.5
oauth2-provider==0.0
oauthlib==1.0.3
pathspec==0.3.4
Pillow==3.3.0
Pygments==2.1.3
python-dateutil==2.5.3
python-openid==2.2.5
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.2
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
